requirements.txt
redis
numpy
Pillow
requests
werkzeug
configparser
opencv-python-headless
flask-restplus==0.12.1
flask==1.0.3